% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_2.R
\name{convert_pie_data}
\alias{convert_pie_data}
\title{convert_pie_data}
\usage{
convert_pie_data(data, col1, col2)
}
\arguments{
\item{data}{raw data from API}

\item{col1}{column 1}

\item{col2}{column 2}
}
\value{
data in the correct format for plotly pie chart
}
\description{
convert data into correct format for plotly pie chart
}
\examples{
\dontrun{convert_pie_data(data, columns)}
}
