% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_1.R
\name{create_figure_1}
\alias{create_figure_1}
\title{Create Figure 1}
\usage{
create_figure_1(data, year_to_date, save = TRUE)
}
\arguments{
\item{data}{raw data from API call}

\item{year_to_date}{Last month in quarter: March, June, September, December}

\item{save}{Whether to save the plot or not. TRUE by default}
}
\value{
year to data sums to plot figure 1
}
\description{
This function produces Figure 1 from the publication.
Plots total number of Nino Registrations from EU, outside the EU and the total.
Assumes data is of type list by default and data frame is the first element of the list.
}
\examples{
\dontrun{create_figure_1(data, year_to_date)}
}
