% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_3.R
\name{create_figure_3}
\alias{create_figure_3}
\title{Create Figure 3}
\usage{
create_figure_3(data, year_to_date)
}
\arguments{
\item{data}{data to plot}

\item{year_to_date}{Last month in quarter: March, June, September, December}
}
\value{
year to data sums to plot figure 1
}
\description{
This function produces Figure 3 from the publication.
Plots total number of Nino Registrations from EU, outside the EU and the total.
Assumes data is of type list by default and data frame is the first element of the list.
}
\examples{
\dontrun{create_figure_1(data, year_to_date)}
}
