% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_sequence_hashmap.R
\docType{data}
\name{date_sequence}
\alias{date_sequence}
\title{Creates hashmap containing sequence of numbers with keys as the dates
This helps generate the apicall to match the correct time periods
chosen with the quarter number used by the api call.}
\format{An object of class \code{Date} of length 96.}
\usage{
date_sequence
}
\description{
Creates hashmap containing sequence of numbers with keys as the dates
This helps generate the apicall to match the correct time periods
chosen with the quarter number used by the api call.
}
\keyword{datasets}
