% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_end_data.R
\name{extract_year_end_data}
\alias{extract_year_end_data}
\title{Plots total number of Nino Registrations from EU, outside the EU and the total.
Assumes data is of type list by default and data frame is the first element of the list.}
\usage{
extract_year_end_data(data, year_to_date, year_end_filter, labels)
}
\arguments{
\item{data}{data to plot: generated by year_to_date_sum()}

\item{year_to_date}{one of Mar, Jun, Sep, Dec}

\item{year_end_filter}{column to extract: e.g. quarter_dates}

\item{labels}{list of columns we want to extract}
}
\value{
filtered data for plotting
}
\description{
Plots total number of Nino Registrations from EU, outside the EU and the total.
Assumes data is of type list by default and data frame is the first element of the list.
}
\examples{
\dontrun{extract_year_end_data(data, date_filter, labels)}
}
