% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dates.R
\name{format_dates}
\alias{format_dates}
\title{This function formats dates for year end filtering calculations needed to produce figure 1.}
\usage{
format_dates(data)
}
\arguments{
\item{data}{raw data obtained from the API}
}
\value{
returns the formatted dates vector
}
\description{
This function formats dates for year end filtering calculations needed to produce figure 1.
}
\examples{
\dontrun{format_dates (data)}
}
