% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_2.R
\name{create_figure_2}
\alias{create_figure_2}
\title{create_figure_2}
\usage{
create_figure_2(data, year_to_date)
}
\arguments{
\item{data}{data to plot}

\item{year_to_date}{Last month in quarter: one of Mar, Jun, Sep, Dec}
}
\value{
pie chart showing number of registrations from EU vs Non-EU
}
\description{
This function produces Figure 2 (pie chart) in the Nino Statistical Publication
}
\examples{
\dontrun{create_figure_2(data, year_to_date)}
}
