% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_3.R
\name{create_figure_3}
\alias{create_figure_3}
\title{Create Figure 3}
\usage{
create_figure_3(data, year_to_date, year_end_filter, labels)
}
\arguments{
\item{data}{raw data from API call}

\item{year_to_date}{Last month in quarter, one of: Mar, Jun, Sep, Dec}

\item{year_end_filter}{raw data from API call}

\item{labels}{labels from API call must be in colnames of data
use data$colnames to see available choices}
}
\value{
plot of class ggplot
}
\description{
This function produces Figure 3 from the publication.
Plots total number of Nino Registrations from EU2, EU8 and EU15.
}
\examples{
\dontrun{create_figure_3(data, year_to_date)}
}
