% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_non_eu_total.R
\name{eu_non_eu_total}
\alias{eu_non_eu_total}
\title{Function accesses the Stat-Xplore API and returns data relating
to Nino Registrations from the EU and outside the EU. This data is used
to construct a time series graph of these registrations over time.}
\source{
\url{https://stat-xplore.dwp.gov.uk/webapi/jsf/login.xhtml}
}
\usage{
eu_non_eu_total(start_date, end_date, frequency, get_countries = FALSE,
  verbose = FALSE, apiKey = NULL)
}
\arguments{
\item{start_date}{start date of data we want: month must be one of 3,6,9,12 so "2017-03-01" is valid "2017-01-01" is not}

\item{end_date}{what date do we want data until}

\item{frequency}{One of quarter, month, annual}

\item{get_countries}{if TRUE will get all nationalities, used for top 5 table in publication.}

\item{verbose}{if true, it will return information sent to the Stat-Xplore server.}

\item{apiKey}{apiKey for stat-xplore: you will need to create an account to generate a key}
}
\value{
class containing data and other attributes useful for plotting
}
\description{
Function accesses the Stat-Xplore API and returns data relating
to Nino Registrations from the EU and outside the EU. This data is used
to construct a time series graph of these registrations over time.
}
\examples{
\dontrun{eu_non_eu_total("2002-03-01", "2018-12-01", "quarter", 3, TRUE)}
}
