% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_to_date_sum.R
\name{year_to_date_sum}
\alias{year_to_date_sum}
\title{This function calculates the Year to date sum for Figure 1 in the report}
\usage{
year_to_date_sum(data, year_to_date)
}
\arguments{
\item{data}{raw data obtained from the API}

\item{year_to_date}{Last month in quarter: March, June, September, December}
}
\value{
list containing data frame, dates and length of time series
}
\description{
This function calculates the Year to date sum for Figure 1 in the report
}
\examples{
year_to_date_sum (data, "March")
}
